+++ common/buf/buffer_test.go
@@ -16,10 +16,10 @@ func TestBufferClear(t *testing.T) {
 
 	payload := "Bytes"
 	buffer.Append([]byte(payload))
-	assert(buffer.Len(), Equals, len(payload))
+	assert(buffer.Len(), Equals, int32(len(payload)))
 
 	buffer.Clear()
-	assert(buffer.Len(), Equals, 0)
+	assert(buffer.Len(), Equals, int32(0))
 }
 
 func TestBufferIsEmpty(t *testing.T) {
